Beidou/requirements.txt
alembic==1.0.0
asn1crypto==0.24.0
blinker==1.4
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
Click==7.0
cryptography==2.3.1
cycler==0.10.0
Flask==1.0.2
Flask-Mail==0.9.1
Flask-Migrate==2.2.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
freeze==1.0.10
gunicorn==19.9.0
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
kiwisolver==1.0.1
Mako==1.0.7
MarkupSafe==1.0
matplotlib==3.0.2
numpy==1.15.4
Pillow==5.3.0
pycparser==2.19
PyMySQL==0.9.2
pyparsing==2.3.0
python-dateutil==2.7.3
python-editor==1.0.3
requests==2.20.0
six==1.11.0
SQLAlchemy==1.2.12
urllib3==1.24.1
Werkzeug==0.14.1
